legacy: move legacy AES-CTR key classes into borg.legacy.crypto, refs #9556
